chore: improve test coverage and add missing stories
