+++ src/components/CurriculumComponents/CurricSuccessMessage/SuccessMessage.stories.tsx
