+++ src/components/CurriculumComponents/CurricSuccessMessage/SuccessMessage.test.tsx
