+++ src/components/CurriculumComponents/CurricSuccessMessage/index.tsx
