+++ src/components/CurriculumComponents/CurricVisualiserUnitList/CurricVisualiserUnitList.stories.tsx
@@ -1,0 +1,27 @@
+import { OakThemeProvider, oakDefaultTheme } from "@oaknational/oak-components";
+import { Meta, StoryObj } from "@storybook/react";
+
+import { basicFixtures } from "./CurricVisualiserUnitList.fixtures";
+
+import { CurricVisualiserUnitList as Component } from ".";
+
+const meta: Meta<typeof Component> = {
+  component: Component,
+  argTypes: {},
+};
+
+export default meta;
+type Story = StoryObj<typeof Component>;
+
+export const CurricVisualiserUnitList: Story = {
+  args: {
+    ...basicFixtures,
+  },
+  render: function Render(args) {
+    return (
+      <OakThemeProvider theme={oakDefaultTheme}>
+        <Component {...args} />
+      </OakThemeProvider>
+    );
+  },
+};
